==> code/02_Data Cleaning Macro Level.Rmd <==
---
title: "01.2_Macro Data Prep"
output:
  html_document:
    df_print: paged
---

```{r setup}
pacman::p_load("tidyverse", 
               "countrycode",
               "readstata13")

knitr::opts_chunk$set(message = F, warning = F)
```


## Load

The data cleaning is done in the file "01.1_Data Cleaning Individual Level.Rmd" and country-data prepped in folder [prep country level data](../data/prep country level data/).

```{r loaddata, message = F, warning = F}
# ISSP Data
load(here::here("data","socstat.Rda"))

# create unique wave-year identifier
df <- df %>%
  mutate(iso3c_wave = paste(iso3c,wave, sep = "_"))

# remove DV NA
df_na <- df %>%
  subset(!is.na(incdiff) & !is.na(inclarge))

# check that data are correct
# df_sum <- df %>% count(cyear)
# Yes, wave counts look good

# Social spending data
socx <- read_csv(here::here("data","socx_oecd.csv"))

# Gini data
gini <- read_csv(here::here("data","gini_solt.csv"))

# WVS/IVS data aggregated
wvs <- readRDS(here::here("data","ivs_aggregated.rds")) %>%
  mutate(iso3c = countrycode(COW_NUM, "cown", "iso3c"),
         wave = case_when(
           year %in% c(1980:1989) ~ 1987,
           year %in% c(1990:1995) ~ 1992,
           year %in% c(1996:2005) ~ 1999,
           year %in% c(2006:2015) ~ 2009,
           year %in% c(2016:2020) ~ 2019,
           TRUE ~ NA_real_),
         iso3c_wave = paste(iso3c,wave, sep = "_")) # fit years of WVS into nearest wave of ISSP 1987, 1992, 1999, 2009, 2019
```

### GDP/Regime/Immigration

```{r loaddata2, message = F, warning = F}

# Institutional Logic - welfare state penetration
# for future analyses (social policy variables?)

# GDP and Democracy v Authoritarian from V-Dem (some interpolation, see Breznau and Lanver 2021)
gdp <- read_csv(here::here("data","vdem_gdp_pc_INTERPOLATED_consolidated.csv")) %>%
  mutate(iso3c = iso3) %>%
  dplyr::select(-c(iso3, cow_code, country_name))
regime <- read_csv(here::here("data","vdem_regime__INTERPOLATED_consolidated.csv")) %>%
  mutate(iso3c = iso3) %>%
  dplyr::select(-c(iso3, cow_code, country_name))

regime <- regime %>% # impute means by country
  group_by(iso3c) %>%
  mutate(regime_ix = mean(regime, na.rm = T),
         regime_i = ifelse(is.na(regime), regime_ix, regime)) %>%
  ungroup() %>%
  dplyr::select(-regime_ix)
  
# Corruption perceptions
crpt <- readRDS(here::here("data","cpi.RDS"))
cpi_long <- read_csv(here::here("data", "cpi_long.csv"))

# Immigration
immigration_wb <- read_csv(here::here("data","API_SM.POP.TOTL.ZS_DS2_en_csv_v2_3363225.csv"), skip = 4, col_select = -c("Country Name","Indicator Name","Indicator Code"))

immigration_wb <- immigration_wb %>%
  mutate(iso3c = `Country Code`) %>%
  dplyr::select(-`Country Code`)

colnames(immigration_wb)[1:61] <- paste0("pct_fb_W",colnames(immigration_wb)[1:61])

imm_long <- reshape(as.data.frame(immigration_wb), sep = "_W", times = c(1960:2020), idvar = "iso3c", direction = "long", varying = list(1:61), timevar = "year")

# fill in average scores between observations
imm_long <- imm_long[2:4]
imm_long <- imm_long %>%
  group_by(iso3c) %>%
  mutate(pct_fb = ifelse(year == 1960, 0, pct_fb_W1960),
         pct_fb = ifelse(year == 2015 & is.na(pct_fb), lag(pct_fb, 5), pct_fb),
         pct_fb = ifelse(year == 2020, lag(pct_fb, 5), pct_fb),
         pct_fb_front = zoo::na.locf(pct_fb, fromLast = F, na.rm = T),
         pct_fb_back = zoo::na.locf(pct_fb, fromLast = T, na.rm = T)) %>%
  rowwise() %>%
  mutate(pct_fb_i = mean(c(pct_fb_back,pct_fb_front), na.rm = T)) %>%
  ungroup() %>%
  dplyr::select(iso3c, year, pct_fb_i)
```

### SOCX/Gini

```{r loaddata3, message = F, warning = F}
# Take averages of social spending and Gini to get the most possible countries

socx_20 <- socx %>%
  subset(year > 2015) %>%
  group_by(iso3c) %>%
  summarise_all(mean, na.rm = T) %>%
  mutate(year = 2020)

gini_19 <- gini %>%
  subset(year == 2017) %>%
  mutate(year = 2019)

gini_20 <- gini %>%
  subset(year == 2017) %>%
  mutate(year = 2020)

socx <- rbind(socx, socx_20)
gini <- rbind(gini, gini_19, gini_20)


```

## Add in Role of Gov

From: Breznau, Nate. 2019. Underlying Attitude...



```{r addrog}
rog <- readstata13::read.dta13(here::here("data", "gintijs.dta")) %>%
  mutate(countryx = ifelse(countryx == 278, 276, countryx),
         iso3c = countrycode::countrycode(countryx, "iso3n", "iso3c")) %>%
  subset(!is.na(iso3c), select = c(iso3c, year, gint, gintsd))
  
```

### Merges

```{r merge, message = F, warning = F}

# first merges

df_na <- left_join(df_na, gini, by = c("iso3c","year"))
df_na <- left_join(df_na, socx, by = c("iso3c","year"))
df_na <- left_join(df_na, imm_long, by = c("iso3c","year"))
df_na <- left_join(df_na, crpt, by = c("iso3c_wave"))

rog <- left_join(rog, gini, by = c("iso3c","year"))
rog <- left_join(rog, socx, by = c("iso3c","year"))
rog <- left_join(rog, imm_long, by = c("iso3c","year"))

rog <- rog %>%
  mutate(iso3c_wave = paste(iso3c,year, sep = "_"))

rog <- left_join(rog, cpi_long, by = c("iso3c", "year"))

# the WVS/IVS has repeated years, collapse them

wvs_2 <- wvs %>%
  group_by(iso3c_wave) %>%
  summarise_all(mean, na.rm = T) %>%
  dplyr::select(iso3c_wave, liberal, liberal_redist, confidence_gov)


df_na <- left_join(df_na, wvs_2, by = c("iso3c_wave"))
df_na <- left_join(df_na, gdp, by = c("iso3c", "year"))
df_na <- left_join(df_na, regime, by = c("iso3c", "year"))

rog <- left_join(rog, gdp, by = c("iso3c", "year"))
df_macro_rog <- left_join(rog, regime, by = c("iso3c", "year"))

# create an interpolated version of wvs variables that takes the mean by country to fill in missing - most countries do not change much over time

# get country means
df_na <- df_na %>% 
  group_by(iso3c) %>%
  mutate(liberal_ix = mean(liberal, na.rm = T),
         liberal_redist_ix = mean(liberal_redist, na.rm = T),
         confidence_gov_ix = mean(confidence_gov, na.rm = T)) %>%
  ungroup()

# replace missing with means
df_na <- df_na %>%
  mutate(liberal_i = ifelse(is.na(liberal), liberal_ix, liberal),
         liberal_redist_i = ifelse(is.na(liberal_redist), liberal_redist_ix, liberal_redist),
         confidence_gov_i = ifelse(is.na(confidence_gov), confidence_gov_ix, confidence_gov),
         liberal = ifelse(liberal == "NaN", NA, liberal),
         liberal_redist = ifelse(liberal_redist == "NaN", NA, liberal_redist),
         confidence_gov = ifelse(confidence_gov == "NaN", NA, confidence_gov))


rm(gini,socx,crpt, wvs_2)
rm(imm_long)
```

### Fix Missing

Where possible

```{r wrangle}
# make a complete data frame
df_na <- df_na %>%
  group_by(iso3c) %>%
  # calculate the mean for each country
  mutate(incdiff_b = mean(incdiff, na.rm = T),
         inclarge_b = mean(inclarge, na.rm = T),
         reduce_b = mean(reduce, na.rm = T),
         reduce_large_b = mean(reduce_large, na.rm = T)) %>%
  ungroup() %>%
  # calculate the mean-centered values by subtracting the means
  mutate(incdiff_w = incdiff - incdiff_b,
         inclarge_w = inclarge - inclarge_b,
         reduce_w = reduce - reduce_b,
         reduce_large_w = reduce_large - reduce_large_b,
         democratic = regime)


```


#### Identify Missing

There is some missing data
identify and fill in

```{r missing_wrangle}
df_na %>%
  dplyr::select(iso3c, year, socx, gini_disp) %>%
  group_by(iso3c, year) %>%
  summarise_all(mean, na.rm = T)


```


#### Create Macro-Level Data

```{r saveout}
# The variables names are not intuitive, change them

df_na <- df_na %>%
  rename(gov_redist = reduce, incdiff_large = reduce_large)

# create a variable for the percentage of the population that agrees income gap is too large, but does not agree govt is responsible to reduce it
#plus a percent agree with gov redist variable
df_na <- df_na %>%
  mutate(gap_no_gov = ifelse(incdiff_large > 3 & gov_redist < 4, 100, 0),
         gov_redist_agree = ifelse(gov_redist > 3, 100, ifelse(gov_redist < 3, 0, NA)),
         incdiff_large_agree = ifelse(incdiff_large > 3, 100, ifelse(incdiff_large < 3, 0, NA)),
         educyrs = ifelse(educyrs < 0, NA, educyrs),
         pct_sample_tertiary = ifelse(educyrs > 15, 100, 0), 
         pct_sample_65older = ifelse(ageC > 64, 100, 0))

# create a macro-level dataset
df_macro <- df_na %>%
  dplyr::select(iso3c, wave, gov_redist, incdiff_large, gini_disp, gini_mkt, socx, cpi, cpi_ii, liberal, liberal_i, liberal_redist, liberal_redist_i, confidence_gov, confidence_gov_i, democratic, filled_vdem_gdp_interpol, regime, gap_no_gov, gov_redist_agree, incdiff_large_agree, pct_sample_tertiary, pct_sample_65older, pct_fb_i) %>%
  group_by(iso3c, wave) %>%
  summarise_all(mean, na.rm = T) %>%
  ungroup()

df_macro[df_macro == "NaN"] <- NA


#fill in non-OECD by hand
# https://www.social-protection.org/gimi/WSPDB.action?id=19
```

#### Replace Missing

```{r saveout2}
df_macro <- df_macro %>%
  mutate(socx = ifelse(iso3c == "AUT" & wave == 1987, 24, socx),
         socx = ifelse(iso3c == "BGR" & wave == 1993, 25, socx),
         socx = ifelse(iso3c == "BGR" & wave == 1999, 38, socx),
         socx = ifelse(iso3c == "BGR" & wave == 2009, 40, socx),
         socx = ifelse(iso3c == "CHE" & wave == 1987, 10, socx),
         socx = ifelse(iso3c == "CHE" & wave == 2009, 22, socx),
         socx = ifelse(iso3c == "CHE" & wave == 2019, 20, socx),
         socx = ifelse(iso3c == "CYP" & wave == 1999, 18, socx),
         socx = ifelse(iso3c == "CYP" & wave == 2009, 25, socx),
         socx = ifelse(iso3c == "HRV" & wave == 2019, 30, socx),
         socx = ifelse(iso3c == "HUN" & wave == 1987, 20, socx),
         socx = ifelse(iso3c == "HUN" & wave == 1992, 35, socx),
         socx = ifelse(iso3c == "HUN" & wave == 1998, 40, socx),
         socx = ifelse(iso3c == "JPN" & wave == 2019, 25, socx),
         socx = ifelse(iso3c == "PHL" & wave == 1992, 2, socx),
         socx = ifelse(iso3c == "PHL" & wave == 1999, 2.5, socx),
         socx = ifelse(iso3c == "PHL" & wave == 2009, 3, socx),
         socx = ifelse(iso3c == "PHL" & wave == 2019, 4, socx),
         socx = ifelse(iso3c == "POL" & wave == 1987, 25, socx),
         socx = ifelse(iso3c == "RUS" & wave == 1992, 25, socx),
         socx = ifelse(iso3c == "RUS" & wave == 1999, 23, socx),
         socx = ifelse(iso3c == "RUS" & wave == 2009, 21.5, socx),
         socx = ifelse(iso3c == "RUS" & wave == 2019, 20, socx),
         socx = ifelse(iso3c == "SVK" & wave == 1992, 20, socx),
         socx = ifelse(iso3c == "SVN" & wave == 1992, 25, socx),
         socx = ifelse(iso3c == "ZAF" & wave == 2019, 25, socx),
         socx = ifelse(iso3c == "ZAF" & wave == 2019, 25, socx),
         socx = ifelse(iso3c == "THA" & wave == 2019, 6, socx),
         gini_disp = ifelse(iso3c == "PHL" & wave == 2019, 42, gini_disp),
         gini_disp = ifelse(iso3c == "PHL" & wave == 2019, 42, gini_disp),
         gini_disp = ifelse(iso3c == "JPN" & wave == 2019, 35, gini_disp),
         gini_disp = ifelse(iso3c == "ZAF" & wave == 2019, 50, gini_disp),
         gini_disp = ifelse(iso3c == "ZAF" & wave == 2019, 50, gini_disp),
         gdp_pc_10k = filled_vdem_gdp_interpol/10000, # make a GDP 10k var
         gdp_pc_10k = ifelse(iso3c == "CHE" & wave == 2019, 5.6, gdp_pc_10k), # google gdps
         gdp_pc_10k = ifelse(iso3c == "CHL" & wave == 2019, 1.75, gdp_pc_10k),
         gdp_pc_10k = ifelse(iso3c == "CZE" & wave == 2019, 2.5, gdp_pc_10k),
         gdp_pc_10k = ifelse(iso3c == "DEU" & wave == 2019, 4.5, gdp_pc_10k),
         gdp_pc_10k = ifelse(iso3c == "FIN" & wave == 2019, 4.9, gdp_pc_10k),
         gdp_pc_10k = ifelse(iso3c == "HRV" & wave == 2019, 1.4, gdp_pc_10k),
         gdp_pc_10k = ifelse(iso3c == "ITA" & wave == 2019, 3.2, gdp_pc_10k),
         gdp_pc_10k = ifelse(iso3c == "JPN" & wave == 2019, 4, gdp_pc_10k),
         gdp_pc_10k = ifelse(iso3c == "NZL" & wave == 2019, 4, gdp_pc_10k),
         gdp_pc_10k = ifelse(iso3c == "PHL" & wave == 2019, 0.8, gdp_pc_10k),
         gdp_pc_10k = ifelse(iso3c == "RUS" & wave == 2019, 1.8, gdp_pc_10k),
         gdp_pc_10k = ifelse(iso3c == "SVN" & wave == 2019, 2.6, gdp_pc_10k),
         gdp_pc_10k = ifelse(iso3c == "THA" & wave == 2019, 0.9, gdp_pc_10k),
         socx = ifelse(iso3c == "BGR" & wave == 1992, 38, socx),
         socx = ifelse(iso3c == "HUN" & wave == 1999, 30, socx),
         gdp_pc_10k = ifelse(iso3c == "DNK" & wave == 2019, 5, gdp_pc_10k))

# there is some mismatch when summarizing, fix ivs variables

df_macro <- df_macro %>%
  mutate(liberal_i = ifelse(is.na(liberal), liberal_i, liberal),
         confidence_gov_i = ifelse(is.na(confidence_gov), confidence_gov_i, confidence_gov),
         liberal_redist_i = ifelse(is.na(liberal_redist), liberal_redist_i, liberal_redist))


```

### Correlations

```{r figx}
# get all correlations

c_corrs = df_na %>% # have to use equal sign here because it is a function
  subset(!is.na(gov_redist) & !is.na(incdiff_large)) %>% # remove missing cases
  group_by(iso3c, wave) %>% # group by country-wave
  mutate(gov_redist_Z = scale(gov_redist),
         incdiff_large_Z = scale(incdiff_large)) %>%
  do(model = lm(gov_redist_Z ~ incdiff_large_Z, data = .)) #run regression for each country

for (j in 1:102) {
c_corrs$corr[j] <- c_corrs[[3]][[j]][["coefficients"]][2]
}

c_corrs <- dplyr::select(c_corrs, -model)
# merge with df_macro
df_macro <- left_join(df_macro, c_corrs, by = c("iso3c", "wave"))

# for plotting lines of best fit


# first fill in a few missing gdp
df_macro <- df_macro %>%
  group_by(iso3c) %>%
  mutate(gdp_mean = mean(gdp_pc_10k, na.rm = T),
         gdp_i = ifelse(is.na(gdp_pc_10k), gdp_mean, gdp_pc_10k)) %>%
  ungroup()

# create a variable with all values in one, and then split by high and middle income countries
# need to find a way to adjust for economic growth and unblanaced time points for calculating GDP
rich_countries <- c("NOR","CHE","DNK","USA","AUS","DEU","AUT","SWE","JPN","NZL","GBR","FRA","CAN","ITA","FIN","NOR","IRL","NLD")
df_macro <- df_macro %>%
  mutate(liberal_high = ifelse(iso3c %in% rich_countries, liberal, NA), 
         liberal_low = ifelse(is.na(liberal_high), liberal, NA),
         liberal_low = ifelse(iso3c %in% c(
                                           "SVN",
                                           "CZE","CYP", "ISR"
                                           #,"ESP","PRT"
                                           ), NA, liberal_low), # remove some questionable mid-GDP cases
         liberal_high_i = ifelse(iso3c %in% rich_countries, liberal_i, NA), 
         liberal_low_i = ifelse(is.na(liberal_high), liberal_i, NA),
         liberal_low_i = ifelse(iso3c %in% c(
                                             "SVN",
                                             "CZE","CYP", "ISR"
                                             #,"ESP","PRT"
                                             ), NA, liberal_low_i),
         liberal_redist_high = ifelse(iso3c %in% rich_countries, liberal_redist, NA), 
         liberal_redist_low = ifelse(is.na(liberal_redist_high), liberal_redist, NA),
         liberal_redist_low = ifelse(iso3c %in% c(
                                                  "SVN",
                                                  "CZE","CYP","ISR"
                                                  #,"ESP","PRT"
                                                  ), NA, liberal_redist_low), # remove some questionable mid-GDP cases
         liberal_redist_high_i = ifelse(iso3c %in% rich_countries, liberal_redist_i, NA), 
         liberal_redist_low_i = ifelse(is.na(liberal_redist_high), liberal_redist_i, NA),
         liberal_redist_low_i = ifelse(iso3c %in% c(
                                                    "SVN",
                                                    "CZE","CYP", "ISR"
                                                    #,"ESP","PRT"
                                                    ), NA, liberal_redist_low_i),
         gdp_di = ifelse(!is.na(liberal_low_i), 0, 1)) 



df_macro <- df_macro %>%
  mutate(cpi_ii = ifelse(iso3c == "DNK", 90, cpi_ii))


df_macro <- df_macro %>%
  mutate(cpiZ_i = -1*as.numeric(scale(cpi_ii)), # make higher scores = more corrupt percept
         cpiZ = -1*as.numeric(scale(cpi)),
         libZ_i = as.numeric(scale(liberal_i)),
         libZ = as.numeric(scale(liberal)),
         lib_redistZ_i = as.numeric(scale(liberal_redist_i)),
         lib_redistZ = as.numeric(scale(liberal_redist)),
         noconf_govZ_i = -1*as.numeric(scale(confidence_gov_i)),
         noconf_govZ = -1*as.numeric(scale(confidence_gov)))

# make high and low corruption perceptions and noconf
df_macro <- df_macro %>%
  mutate(noconf_govZ_i = ifelse(noconf_govZ_i < -2, -2, noconf_govZ_i), # too far outliers
         noconf_govZ_i = ifelse(noconf_govZ_i > 2, 2, noconf_govZ_i),
         noconf_govZ = ifelse(noconf_govZ < -2, -2, noconf_govZ),
         noconf_govZ = ifelse(noconf_govZ > 2, 2, noconf_govZ),
         cpi_low = ifelse(!is.na(liberal_low_i), cpiZ, NA),
         cpi_high = ifelse(!is.na(liberal_high_i), cpiZ, NA),
         noconf_govZ_low = ifelse(!is.na(liberal_low_i), noconf_govZ, NA),
         noconf_govZ_hi = ifelse(!is.na(liberal_high_i), noconf_govZ, NA),
         gdp_lo = ifelse(!is.na(liberal_low_i), gdp_pc_10k, NA),
         gdp_hi = ifelse(!is.na(liberal_high_i), gdp_pc_10k, NA)) %>%
  dplyr::select(iso3c, wave, gov_redist, incdiff_large, corr, gini_disp, gini_mkt, socx, gdp_pc_10k, pct_fb_i, liberal, liberal_i, liberal_high, liberal_low, liberal_high_i, liberal_low_i, cpi_high, cpi_low, noconf_govZ_low, noconf_govZ_hi, gdp_lo, gdp_hi, gdp_di, liberal_redist, liberal_redist_i, liberal_redist_high, liberal_redist_low, liberal_redist_high_i, liberal_redist_low_i, confidence_gov, confidence_gov_i, cpi, cpi_ii, noconf_govZ, noconf_govZ_i, cpiZ, cpiZ_i, libZ, libZ_i, lib_redistZ, lib_redistZ_i, gap_no_gov, gov_redist_agree, incdiff_large_agree)


# create within and between values
df_na <- df_na %>%
  group_by(iso3c,wave) %>%
  mutate(incdiff_large_b = mean(incdiff_large, na.rm = T)) %>%
  ungroup()

df_na <- df_na %>%
  mutate(incdiff_large_w = incdiff_large - incdiff_large_b,
         iso3c_wave = paste(iso3c,wave, sep = "_"))

```

## Final Clean Up

```{r prep_data}
df_na <- df_na %>%
  dplyr::select(-c(gini_disp, socx)) 

df_macro <- as.data.frame(df_macro)

df_macro_m <- df_macro %>%
  dplyr::select(iso3c, wave, cpiZ, cpiZ_i, libZ, libZ_i, lib_redistZ, lib_redistZ_i, gdp_di, gdp_pc_10k, gini_disp, socx, noconf_govZ, noconf_govZ_i)

# check that all macro variables are centered, ideal for MLM
df_macro_m$gdp_pc_10k_C <- df_macro_m$gdp_pc_10k - mean(df_macro_m$gdp_pc_10k, na.rm = T)
df_macro_m$gini_disp_C <- df_macro_m$gini_disp - mean(df_macro_m$gini_disp, na.rm = T)
df_macro_m$socx_C <- df_macro_m$socx - mean(df_macro_m$socx, na.rm = T)


df_na <- df_na %>%
  left_join(df_macro_m, by = c("iso3c", "wave"))

df_na$incdiff_large_C <- df_na$incdiff_large - mean(df_na$incdiff_large, na.rm = T)
df_na$gov_redist_C <- df_na$gov_redist - mean(df_na$gov_redist, na.rm = T)


df_na$pct_fb_i_C <- df_na$pct_fb_i - mean(df_na$pct_fb_i, na.rm = T)
```


## Add Social Security Data

Introduction and Coverage

```{r mainiv}
# work injury 
work_inj <- read_csv(here::here("data", "gwip_v1.csv"))

# work injury coverage
eiip <- read_csv(here::here("data", "EIIP_2014.csv"))

# unemp
unemp <- read_csv(here::here("data", "labor_unemp_firstlaw.csv")) %>%
  mutate(iso3c = countrycode(cow_code, "cown", "iso3c"),
         unemp_first_law = as.numeric(year)) %>%
  dplyr::select(iso3c, unemp_first_law)

# pensions (take US social security admin scraped because WeSIS data is not all countries, see scraping folder)
pension <- readRDS(here::here("data", "ws_intro.rds")) %>%
  mutate(iso3c = countrycode(cow_code, "cown", "iso3c"),
         pension_first_law = as.numeric(old_pension_firstlaw)) %>%
  dplyr::select(iso3c, pension_first_law)

work_inj <- work_inj %>%
  mutate(iso3c = countrycode(cow_code, "cown", "iso3c"))

eiip <- eiip %>%
  mutate(iso3c = countrycode(Country, "country.name", "iso3c"),
         winj_cov = as.numeric(Coverage_pct_LF)) %>%
  dplyr::select(iso3c, winj_cov)

df_na <- left_join(df_na, work_inj, by = "iso3c")
df_na <- left_join(df_na, eiip, by = "iso3c")
df_na <- left_join(df_na, unemp, by = "iso3c")
df_na <- left_join(df_na, pension, by = "iso3c")

df_macro <- left_join(df_macro, work_inj, by = "iso3c")
df_macro <- left_join(df_macro, eiip, by = "iso3c")
df_macro <- left_join(df_macro, unemp, by = "iso3c")
df_macro <- left_join(df_macro, pension, by = "iso3c")

df_macro_rog <- left_join(df_macro_rog, work_inj, by = "iso3c")
df_macro_rog <- left_join(df_macro_rog, eiip, by = "iso3c")
df_macro_rog <- left_join(df_macro_rog, unemp, by = "iso3c")
df_macro_rog <- left_join(df_macro_rog, pension, by = "iso3c")
```

## SOCX for the ROG dataset
```{r socxrog}
df_macro_rog <- df_macro_rog %>%
  mutate(socx = case_when(
    iso3c == "AUT" ~ 22,
    iso3c == "HRV" ~ 30,
    iso3c == "BGR" ~ 40,
    iso3c == "CYP" ~ 20,
    iso3c == "DOM" ~ 10,
    iso3c == "HUN" & year == "1990" ~ 34,
    iso3c == "HUN" & year == "1996" ~ 37,
    iso3c == "IND" ~ 7.5,
    iso3c == "PHL" & year == "1996" ~ 3,
    iso3c == "PHL" & year == "2006" ~ 4,
    iso3c == "PHL" & year == "2016" ~ 6,
    iso3c == "RUS" & year == "1996" ~ 27,
    iso3c == "RUS" & year == "2006" ~ 25,
    iso3c == "RUS" & year == "2016" ~ 23,
    iso3c == "ZAF" & year == "2006" ~ 25,
    iso3c == "ZAF" & year == "2016" ~ 25,
    iso3c == "SUR" ~ 12,
    iso3c == "THA" ~ 3,
    iso3c == "URY" ~ 21,
    iso3c == "VEN" ~ 18,
    iso3c == "ISR" & year == "1990" ~ 15,
    TRUE ~ socx
  ))
```

## Add Typology

```{r types}
df_macro <- df_macro %>%
  mutate(type_7 = case_when(
    iso3c %in% c("AUS", "NZL") ~ "G. Radical",
    iso3c %in% c("USA", "GBR", "IRL", "JPN", "CHE", "CAN") ~ "E. Liberal",
    iso3c %in% c("AUT", "BEL", "DEU", "FRA", "POL") ~ "B. Conservative European",
    iso3c %in% c("NLD", "DNK", "SWE", "NOR", "FIN") ~ "C. Social Dem",
    iso3c %in% c("RUS", "BGR", "HRV", "HUN", "LVA", "SVK", "SVN", "CZE") ~ "D. Former Communist",
    iso3c %in% c("PHL", "THA", "ZAF", "ISR", "CHL") ~ "A. Uncategorized",
    iso3c %in% c("CYP", "ITA", "ESP", "PRT") ~ "F. Southern European"
  ))

df_macro_rog <- df_macro_rog %>%
  mutate(type_8 = case_when(
    iso3c %in% c("AUS", "NZL") ~ "G. Radical",
    iso3c %in% c("USA", "GBR", "IRL", "JPN", "CHE", "CAN", "TWN", "KOR") ~ "H. Liberal",
    iso3c %in% c("AUT", "BEL", "DEU", "FRA", "POL") ~ "B. Conservative European",
    iso3c %in% c("NLD", "DNK", "SWE", "NOR", "FIN", "ISL") ~ "C. Social Dem",
    iso3c %in% c("RUS", "BGR", "HRV", "HUN", "LVA", "SVK", "SVN", "CZE", "LTU") ~ "D. Former Communist",
    iso3c %in% c("THA", "PHL", "ZAF", "ISR", "IND", "SUR", "TUR") ~ "A. Uncategorized",
    iso3c %in% c("DOM", "CHL", "URY", "VEN") ~ "E. Latin American",
    iso3c %in% c("CYP", "ITA", "ESP", "PRT") ~ "F. Southern European"
  ))
```

## Institutions in Years

```{r instyear}
df_na <- df_na %>%
  mutate(soc_sec_age = (year - unemp_first_law) + (year - pension_first_law) + (year - labor_workinjury_firstlaw))

# add year for macro
df_macro <- df_macro %>%
  mutate(soc_sec_age = (wave - unemp_first_law) + (wave - pension_first_law) + (wave - labor_workinjury_firstlaw))

df_macro_rog <- df_macro_rog %>%
  mutate(soc_sec_age = (year - unemp_first_law) + (year - pension_first_law) + (year - labor_workinjury_firstlaw))
```

## Adjust for SOCX

The problem with SOCX is that it increases over time, so it is necessary to take a cross-sectional version. Take the mean by country. This is imperfect because it is an unbalanced time-series, but it should be good enough for now. The real problem is within-country change which this solves. It also keeps countries relatively sorted in order of 'size of welfare state'. I also do this for GDP here.

```{r socx_cs}
df_macro <- as.data.frame(df_macro) %>%
  group_by(iso3c) %>% 
  mutate(socx_C = mean(socx),
         gdp_C = mean(gdp_pc_10k)) %>% 
  ungroup()

df_macro_rog <- as.data.frame(df_macro_rog) %>%
  group_by(iso3c) %>% 
  mutate(socx_C = mean(socx)) %>% 
  ungroup()
```

### Final Variable Mods

Dummy for former communist
Re-scale for easier regression results
Plot_summs does not standardize the variables properly, so scale them here
```{r fmr_comm}
df_macro <- df_macro %>% 
  mutate(fcomm = ifelse(type_7 == "Former Communist", 1, 0),
         soc_sec_age_100 = soc_sec_age/100,
         w = 6) %>% 
  group_by(iso3c) %>% 
  mutate(groupn = n()) %>% 
  ungroup() %>% 
  mutate(weight = w/groupn)

df_macro_rog <- df_macro_rog %>% 
  mutate(fcomm = ifelse(type_8 == "Former Communist", 1, 0),
         soc_sec_age_100 = soc_sec_age/100,
         w = 6) %>% 
  group_by(iso3c) %>% 
  mutate(groupn = n()) %>% 
  ungroup() %>% 
  mutate(weight = w/groupn)

# count countries, generate weights

```

## Save DF
```{r save}

# with missing cases removed
saveRDS(df_na, file = here::here("data","df_na.RDS"))
# macro only
saveRDS(df_macro, file = here::here("data","df_macro.RDS"))
saveRDS(df_macro_rog, file = here::here("data","df_macro_rog.RDS"))

```

## SIED

Nelson, K., Fredriksson, D., Korpi, T., Korpi, W., Palme, J. and O. Sjöberg. 2020. The Social Policy Indicators (SPIN) database. International Journal of Social Welfare. 29 (3). 285-289. https://doi.org/10.1111/ijsw.12418

### Setup


prtstwco	Pension replacement rate standard worker couple (pbenstwco/gapwyear)
urtsw26f - Standard gross 26-week replacement rate, familied worker (UBESW26F / UGAPWEEK)
artsw26f - Standard gross 26-week replacement rate, familied worker (ABESW26F / AGAPWEEK)

pcovratp - Coverage ratio in population 15-65 years of age
ucovratl - Unemployment insurance coverage ratio as proportion of labour force (UNOINSUR / ULABFORC)
acovratl - Coverage ratio as proportion of labour force (ANOINSUR /ALABFORC)


```{r sied_retr}
sied <- readxl::read_xlsx(here::here("data", "SIED 1930-2015 201012.xlsx")) %>%
  mutate(iso3c = country)  %>%
  select(iso3c, year, prtstwco, urtsw26f, artsw26f, pcovratp, ucovratl, acovratl)

# group by year chunks and remove countries that lack data prior to 1980
sied_plot <- sied %>%
  subset(iso3c %in% unique(df_macro$iso3c) & !(iso3c %in% c("BGR", "CYP", "EST", "GRC", "HUN", "LVA", "LTU", "LUX", "MLT", "PRT", "ROU", "SVK", "SVN", "ESP", "POL"))) %>%
  mutate(year10 = case_when(
    year > 1940 & year < 1955 ~ 1950,
    year >=1955 & year < 1965 ~ 1960,
    year >=1965 & year < 1975 ~ 1970,
    year >=1975 & year < 1980 ~ 1977,
    year >=1980 & year < 1985 ~ 1982,
    year >=1985 & year < 1990 ~ 1987,
    year >=1990 & year < 1995 ~ 1992,
    year >=1995 & year < 2000 ~ 1997,
    year >=2000 & year < 2005 ~ 2002,
    year >=2005 & year < 2010 ~ 2007,
    year >=2010 & year < 2015 ~ 2012,
    year >=2015 ~ 2017
  )) %>%
  subset(!is.na(year10))

sied_plot <- sied_plot %>%
  group_by(year10) %>%
  mutate(pension_rr = mean(prtstwco, na.rm = T),
         unemp_rr = mean(urtsw26f, na.rm = T),
         workinj_rr = mean(artsw26f, na.rm = T),
         pension_cv = mean(pcovratp, na.rm = T),
         unemp_cv = mean(as.numeric(ucovratl), na.rm = T),
         workinj_cv = mean(as.numeric(acovratl), na.rm = T)) %>%
  ungroup()

sied_plot_m <- sied_plot %>%
  group_by(year10) %>%
  dplyr::select(pension_rr:workinj_cv) %>%
  summarise_all(mean, na.rm = T)

saveRDS(sied_plot_m, here::here("data", "sied_plot_m.Rds"))
  
```
## Colophon

```{r colophon}
sessionInfo()
```
