==> code/03 Analysis.Rmd <==
---
title: "03 Analysis"
author: "Nate Breznau"
date: "`r Sys.Date()`"
output: html_document
---



```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

pacman::p_load("tidyverse",
               "ggplot2",
               "ragg",
               "readxl",
               "ggpubr",
               "jtools",
               "ggstance",
               "broom.mixed",
               "QuantPsyc",
               "sjPlot",
               "sjmisc",
               "sjlabelled",
               "lme4")
```

## Data

```{r data}
# with missing cases removed
df_na <- readRDS(here::here("data","df_na.RDS"))
# macro only
df_macro <- readRDS(here::here("data","df_macro.RDS"))
df_macro_rog <- readRDS(here::here("data", "df_macro_rog.RDS"))
sied_plot_m <- readRDS(here::here("data", "sied_plot_m.Rds"))

```





## Fig 1

### Individual Plots

```{r sied_plot}
pension <- sied_plot_m %>%
  ggplot() +
  geom_line(aes(y = pension_rr, x = year10)) +
  geom_line(aes(y = pension_cv, x = year10), linetype = "dashed") +
  ylim(0.2,1) +
  xlim(1950, 2022) +
  labs(x = "", y = "") +
  annotate("text", x = 1956, y = 0.83, label = "Labor Force\nCoverage", hjust = 0) +
  annotate("text", x = 1980, y = 0.51, label = "Wage Replacement", hjust = 0) +
  theme_classic() +
  theme(axis.text.y = element_blank())

unemp <- sied_plot_m %>%
  ggplot() +
  geom_line(aes(y = unemp_rr, x = year10)) +
  geom_line(aes(y = unemp_cv, x = year10), linetype = "dashed") +
  labs(x = "", y = "") +
  ylim(0.2,1) +
  xlim(1950, 2022) +
  theme_classic() +
  theme(axis.text.y = element_blank())

workinj <- sied_plot_m %>%
  ggplot() +
  geom_line(aes(y = workinj_rr, x = year10)) +
  geom_line(aes(y = workinj_cv, x = year10), linetype = "dashed") +
  labs(x = "", y = "Average Rate") +
  ylim(0.2,1) +
  xlim(1950, 2022) +
  theme_classic() 

```
### Arrange Plots

```{r sied_ggarrange}
agg_png(here::here("results", "Fig1.png"), res = 144, width = 1600, height = 600)
ggarrange(workinj, pension, unemp,
          nrow = 1, labels = c("     Work-Injury", "Pensions", "Unemployment"))
dev.off()

knitr::include_graphics(here::here("results", "Fig1.png"))
```






```{r corrs}
# Correlations
cor(dplyr::select(df_macro, gov_redist, liberal_redist_i, socx, socx_C), use = "pairwise.complete.obs")

cor(dplyr::select(df_macro_rog, gint, socx, socx_C), use = "pairwise.complete.obs")

```


```{r socx}

# socx is quite linear w/ gov should redistribute
ggplot(df_macro, aes(y = gov_redist, x = socx)) +
  geom_point()

# socx is somewhat linear w/ inequality is too large
ggplot(df_macro, aes(y = incdiff_large, x = socx)) +
  geom_point()

# socx is quite linear w/ gov intervention scale
ggplot(df_macro_rog, aes(y = gint, x = socx)) +
  geom_point()

# socx is not linear w/ liberal_redist_i
ggplot(df_macro, aes(y = liberal_redist_i, x = socx)) +
  geom_point()

# gdp also linear, but this could include a time trend
ggplot(df_macro, aes(y = gov_redist, x = gdp_pc_10k)) +
  geom_point()

m1 <- lm(gov_redist ~ socx_C, data = df_macro)
m2 <- lm(gint ~ socx_C, data = df_macro_rog)

df_macro <- df_macro %>%
  mutate(gov_redist_socx = predict.lm(m1, newdata = df_macro),
         gov_redist_socx_resid = gov_redist - gov_redist_socx)

df_macro_rog <- df_macro_rog %>%
  mutate(gint_socx = predict.lm(m2, newdata = df_macro_rog),
         gint_socx_resid = gint - gint_socx)


```


## Fig 3 Gov Redist

### Adjusted

```{r plot1}
# check
agg_png(here::here("results", "Fig3_redist.png"), res = 144, height = 600, width = 1000)
ggplot(df_macro, aes(y = gov_redist_socx_resid, x = soc_sec_age)) +
  geom_point(aes(color = type_7, shape = type_7)) +
  scale_color_viridis_d() +
  scale_shape_manual(values = c(15,16,17,8,9,10,11)) +
  geom_smooth(se=FALSE, method="gam", span=0.5) +
  labs(y = "Support for Redistribution", 
       x = "Cumulative Age of Social Policies",
       color = "Regime Type", shape = "Regime Type") +
  theme_classic() 
dev.off()

knitr::include_graphics(here::here("results", "Fig3_redist.png"))
```

### Raw

```{r plot1raw}
# check
agg_png(here::here("results", "Fig3_redist_raw.png"), res = 144, height = 600, width = 1000)
ggplot(df_macro, aes(y = gov_redist, x = soc_sec_age)) +
  geom_point(aes(color = type_7, shape = type_7)) +
  scale_color_viridis_d() +
  scale_shape_manual(values = c(15,16,17,8,9,10,11)) +
  geom_smooth(se=FALSE, method="gam", span=0.5) +
  labs(y = "Support for Redistribution", 
       x = "Cumulative Age of Social Policies",
       color = "Regime Type", shape = "Regime Type") +
  theme_classic() 
dev.off()

knitr::include_graphics(here::here("results", "Fig3_redist_raw.png"))
```
## Fig 4 WS Support

### Adjusted

```{r plot2}
agg_png(here::here("results", "Fig4_wssupport.png"), res = 144, height = 600, width = 1000)
ggplot(df_macro_rog, aes(y = gint_socx_resid, x = soc_sec_age)) +
  geom_point(aes(color = type_8, shape = type_8)) +
  scale_color_viridis_d() +
  scale_shape_manual(values = c(15,16,18,17,8,9,10,11)) +
  geom_smooth(se=FALSE, method="gam", span=0.5) +
  labs(y = "Support for the Welfare State", 
       x = "Cumulative Age of Social Policies",
       color = "Regime Type", shape = "Regime Type") +
  theme_classic()
dev.off()

knitr::include_graphics(here::here("results", "Fig4_wssupport.png"))
```

### Raw

```{r plot2raw}
agg_png(here::here("results", "Fig4_wssupport_raw.png"), res = 144, height = 600, width = 1000)
ggplot(df_macro_rog, aes(y = gint, x = soc_sec_age)) +
  geom_point(aes(color = type_8, shape = type_8)) +
  scale_color_viridis_d() +
  scale_shape_manual(values = c(15,16,18,17,8,9,10,11)) +
  geom_smooth(se=FALSE, method="gam", span=0.5) +
  labs(y = "Support for the Welfare State", 
       x = "Cumulative Age of Social Policies",
       color = "Regime Type", shape = "Regime Type") +
  theme_classic()
dev.off()

knitr::include_graphics(here::here("results", "Fig4_wssupport_raw.png"))
```

## Fig XXX Liberal Values

In an earlier iteration, I also used "Liberal Values" as a 3rd welfare state 'attitude'. However, peer review suggested maybe this was superfluous and I agreed. Therefore, Liberal Values has been removed from the manuscript but left in here in case the reader is interested and for the sake of workflow steps transparency. 

### Adjusted
```{r plot3}
agg_png(here::here("results", "FigX_liberal.png"), res = 144, height = 600, width = 1000)
ggplot(df_macro, aes(y = liberal_redist_i, x = soc_sec_age)) +
  geom_point(aes(color = type_7, shape = type_7)) +
  scale_color_viridis_d() +
  scale_shape_manual(values = c(15,16,17,8,9,10,11)) +
  geom_smooth(se=FALSE, method="gam", span=0.5) +
  labs(y = "Liberal Values", 
       x = "Cumulative Age of Social Policies",
       color = "Regime Type", shape = "Regime Type") +
  theme_classic()
dev.off()

knitr::include_graphics(here::here("results", "FigX_liberal.png"))
```

## Regressions

### Standard

```{r reg}
#without former communist
reg1 <- lm(gov_redist ~ soc_sec_age_100 + socx_C, data = df_macro)
reg2 <- lm(gint ~ soc_sec_age_100 + socx_C, data = df_macro_rog)
reg3 <- lm(liberal_redist_i ~ soc_sec_age_100 + socx_C, data = df_macro)
# standardize
reg1Z <- scale_mod(reg1, scale.response = T)
reg2Z <- scale_mod(reg2, scale.response = T)
reg3Z <- scale_mod(reg3, scale.response = T)
#with
reg4 <- lm(gov_redist ~ soc_sec_age_100 + socx_C + fcomm, data = df_macro)
reg5 <- lm(gint ~ soc_sec_age_100 + socx_C + fcomm, data = df_macro_rog)
reg6 <- lm(liberal_redist_i ~ soc_sec_age_100 + socx_C + fcomm, data = df_macro)
#parabolic
reg5p <- lm(gint ~ soc_sec_age_100 + I(soc_sec_age_100^2) + socx_C + fcomm, data = df_macro_rog)
# standardize
reg4Z <- scale_mod(reg4, scale.response = T)
reg5Z <- scale_mod(reg5, scale.response = T)
reg6Z <- scale_mod(reg6, scale.response = T)
```

### Weighted
```{r regw}
#without former communist
reg1w <- lm(gov_redist ~ soc_sec_age_100 + socx_C, data = df_macro, weights = weight)
reg2w <- lm(gint ~ soc_sec_age_100 + socx_C, data = df_macro_rog, weights = weight) 

#no longer used
reg3w <- lm(liberal_redist_i ~ soc_sec_age_100 + socx_C, data = df_macro, weights = weight)

# standardize
reg1wZ <- scale_mod(reg1w, scale.response = T)
reg2wZ <- scale_mod(reg2w, scale.response = T)
reg3wZ <- scale_mod(reg3w, scale.response = T)
#with
reg4w <- lm(gov_redist ~ soc_sec_age_100 + socx_C + fcomm, data = df_macro, weights = weight)
reg5w <- lm(gint ~ soc_sec_age_100 + socx_C + fcomm, data = df_macro_rog, weights = weight)
#parabolic test
reg5wp <- lm(gint ~ soc_sec_age_100 + I(soc_sec_age_100^2) + socx_C + fcomm, data = df_macro_rog, weights = weight)
reg6w <- lm(liberal_redist_i ~ soc_sec_age_100 + socx_C + fcomm, data = df_macro, weights = weight)
# standardize
reg4wZ <- scale_mod(reg4w, scale.response = T)
reg5wZ <- scale_mod(reg5w, scale.response = T)
reg6wZ <- scale_mod(reg6w, scale.response = T)
```

### Fig 5 Coef plot

#### Standard
Standardize for ease of interpretation
```{r f6a}
agg_png(filename = here::here("results", "Fig5a.png"), res = 144, height = 400, width = 400)
plot_summs(reg4Z, reg5Z,  
           coefs = c("Age of\nSocial Security" = "soc_sec_age_100",
                     "SOCX\nCountry Mean" = "socx_C",
                     "Former\nCommunist" = "fcomm"),
           colors = c("#453781FF", "#73D055FF")) +
           theme(legend.position = "none")

dev.off()

knitr::include_graphics(here::here("results", "Fig5a.png"))

```

```{r f6b}
agg_png(filename = here::here("results", "Fig5b.png"), res = 144, height = 315, width = 550)
plot_summs(reg1Z, reg2Z,  
           coefs = c(" " = "soc_sec_age_100",
                     "  " = "socx_C"),
           model.names = c("Support for\nRedistribution",
                           "Support for\nthe Welfare State"),
           colors = c("#453781FF", "#73D055FF"))

dev.off()

knitr::include_graphics(here::here("results", "Fig5b.png"))
```
### Regression Tables

#### Standard
```{r reg1tables}
tab_model(reg1Z,reg2Z,reg4Z,reg5Z, p.style = "stars", show.ci = F,
          pred.labels = c("(Intercept)",
                          "Age of Social Security",
                          "Social Spending",
                          "Former Communist"),
          dv.labels = c("Support for Redistri- bution",
                        "Support for Welfare State",
                        "Support for Redistri- bution",
                        "Support for Welfare State"),
          file = here::here("results","Table1.doc"))

```

#### Weighted
```{r reg1tablesw}
tab_model(reg1wZ,reg2wZ,reg4wZ,reg5wZ, p.style = "stars", show.ci = F,
          pred.labels = c("(Intercept)",
                          "Age of Social Security",
                          "Social Spending",
                          "Former Communist"),
          dv.labels = c("Support for Redistri- bution",
                        "Support for Welfare State",
                        "Support for Redistri- bution",
                        "Support for Welfare State"),
          file = here::here("results","Table2.doc"))

```

### Fig 6 Predicted Margins

```{r pmargins}
agg_png(here::here("results", "Fig6.png"), res = 144, height = 600, width = 750)

plot_model(reg5, type = "pred", terms = c("soc_sec_age_100", "fcomm")) +
  labs(y = "Support of the Welfare State",
       x = "Age of Social Security",
       title = "",
       color = "Former\nCommunist") +
  scale_color_manual(values = c("#453781FF", "#73D055FF")) +
  annotate("text", x = df_macro_rog$soc_sec_age_100[df_macro_rog$iso3c == "DEU" & df_macro_rog$year == 1996], y = 2.712, hjust = 1, label = "Germany 1996", size = 2.7) +
  annotate("text", x = df_macro_rog$soc_sec_age_100[df_macro_rog$iso3c == "DEU" & df_macro_rog$year == 2016], y = 2.56, hjust = 1, label = "Germany 2016", size = 2.7) +
  annotate("text", x = df_macro_rog$soc_sec_age_100[df_macro_rog$iso3c == "KOR" & df_macro_rog$year == 2006], y = 3.18, hjust = 1, label = "Korea 2006", size = 2.7) +
  annotate("text", x = df_macro_rog$soc_sec_age_100[df_macro_rog$iso3c == "KOR" & df_macro_rog$year == 2016], y = 3.11, hjust = 1, label = "Korea 2016", size = 2.7) +
  annotate("text", x = df_macro_rog$soc_sec_age_100[df_macro_rog$iso3c == "RUS" & df_macro_rog$year == 1996], y = 3.06, hjust = 0, label = "Russia 1996", size = 2.7) +
  annotate("text", x = df_macro_rog$soc_sec_age_100[df_macro_rog$iso3c == "RUS" & df_macro_rog$year == 2016], y = 2.9, hjust = 0, label = "Russia 2016", size = 2.7) +
  annotate("text", x = df_macro_rog$soc_sec_age_100[df_macro_rog$iso3c == "SVN" & df_macro_rog$year == 1996], y = 3.34, hjust = 0, label = "Slovenia 1996", size = 2.7) +
  annotate("text", x = df_macro_rog$soc_sec_age_100[df_macro_rog$iso3c == "SVN" & df_macro_rog$year == 2006], y = 3.27, hjust = 0, label = "Slovenia 2006", size = 2.7) +
  annotate("text", x = df_macro_rog$soc_sec_age_100[df_macro_rog$iso3c == "CAN" & df_macro_rog$year == 1996], y = 2.955, hjust = 1, label = "Canada 1996", size = 2.7) +
  annotate("text", x = df_macro_rog$soc_sec_age_100[df_macro_rog$iso3c == "CAN" & df_macro_rog$year == 2006], y = 2.875, hjust = 1, label = "Canada 2006", size = 2.7) +
  theme_classic() +
  theme(legend.position = c(0.15,0.25))

dev.off()

knitr::include_graphics(here::here("results", "Fig6.png"))
  

```

```{r within_r}
# exclude cases with only one observation
df_macro <- df_macro %>%
  group_by(iso3c) %>%
  mutate(count = n()) %>%
  ungroup()

df_macro_rog <- df_macro_rog %>%
  group_by(iso3c) %>%
  mutate(count = n()) %>%
  ungroup()

grand_corr_redist <- lmer(gov_redist ~ (soc_sec_age_100|iso3c), data = subset(df_macro, count != 1))
grand_corr_gint <- lmer(gint ~ (soc_sec_age_100|iso3c), data = subset(df_macro_rog, count != 1))

#get within-country slopes
redist_slopes <- coef(grand_corr_redist)$iso3c %>%
  mutate(iso3c = rownames(.))
gint_slopes <- coef(grand_corr_gint)$iso3c %>%
  mutate(iso3c = rownames(.))

#add soc_sec_age_100 to slopes data for plotting
avg_age_redist <- df_macro %>%
  dplyr::select(iso3c, soc_sec_age) %>%
  group_by(iso3c) %>%
  summarise(soc_sec_age = mean(soc_sec_age))

avg_age_gint <- df_macro_rog %>%
  dplyr::select(iso3c, soc_sec_age) %>%
  group_by(iso3c) %>%
  summarise(soc_sec_age = mean(soc_sec_age))

redist_slopes <- redist_slopes %>%
  left_join(avg_age_redist, by = "iso3c")

gint_slopes <- gint_slopes %>%
  left_join(avg_age_gint, by = "iso3c")
  
```

# plot within-country slope by age of social security
```{r slope_plot}
agg_png(here::here("results","Fig7.png"), res = 144, height = 800, width = 1000)
ggplot() + 
  geom_text(aes(x = soc_sec_age, y = soc_sec_age_100, label = iso3c, color = "Redistribution"), data = redist_slopes) +
  geom_text(aes(x = soc_sec_age, y = soc_sec_age_100, label = iso3c, color = "Welfare State"), data = gint_slopes) +
  scale_color_manual(values = c("#73D055FF", "#453781FF")) +
  labs(
    y = "Within-Country Slope",
    x = "Average Cumulative Age of Social Security by Country",
    color = "Support for"
  ) +
  theme_classic()

dev.off()

knitr::include_graphics(here::here("results","Fig7.png"))
```
